Throw error on POST /tyk/apis/ if UseDBAppConfigs is enabled
Refer to TykTechnologies/tyk#177

Calling POST to /tyk/apis/ when `UseDBAppConfigs` is enabled may
cause an unexpected behavior. It added a new API definition to
`config.AppPath` directory and didn't add to MongoDB, so it's not
available on the dashboard.

This commit changes the behavior of POST command by rejecting
request if `UseDBAppConfigs` setting is enabled and suggest user to
use an Advanced Management API instead.