Use pointer receivers more consistently
Value receivers are only useful if the value itself needs copying or if
the type is very small and the method is called in a tight loop.

We also modify fields and call other methods in these, so let's err on
the safe side in case any field assignments are being lost.