Merge config_utils.go into config.go
In general, a "utils" file isn't useful. The code should be where it's
used the most. In this particular case, it's config.go, and the whole
thing is independent.

Moreover, the two files together are just over 300 lines. That's not
enough to warrant splitting into multiple files on its own.

Updates #334.