test: implement our own httpbin.org locally
The unit tests depend on online services like httpbin.org. This is not
only flakey and dependant on an internet connection, it also slows down
the tests as on an average connection each roundtrip would be at least
~50ms.

Doing it locally should be safer, easier and faster.

Further commits will start using it.

We can't use httptest because it would randomize the address and urls.
See the comments on testHttpListen.