Set version to v2.3.99
master will become 2.4 at some point, so it should be higher than any
possible 2.3.x version. We're likely not going to have more than a few
bugfix releases, so 2.3.99 is obviously not a stable version.

Don't use a version like v2.4-HEAD or vHEAD as that would break our many
regexes across many repos and configs that expect three numbers.

Ideally, this would include the commit hash of the build. But this is
currently not possible to do with a simple `go get`, so just have it be
an arbitrary but obviously non-stable version.

This removes the confusion of Tyk reporting "version 2.3.2" on master,
which isn't true.