test: fix TestGroupResetHandler pub/sub flakiness
The problem was that if redis was slow (like on Travis), the SUBSCRIBE
command might take a bit of time and we might end up doing the API call
before it's done. If that happened, we never received the redis message
we were expecting.

To fix this, make the callbak also be called for redis.Subscription
events, which mean "pub/sub handler now working". Only after that is it
safe to assume that we will for sure receive our message.

A couple of places in our prod code also don't check for the
subscription to be done, but let's not fix that unless we know that it
might cause issues.