Add unit tests for hot reloads
A test that fires a single hot reload and checks it did something, and
another that fires many concurrently and checks that only a few actually
ran.

Also rework the "hot reload is done" event mechanism. Use a per-call
callback func, which makes the code simpler and use less globals. This
also means that we don't have to worry about hot reloads fired by other
tests.