Use a ticker for hot reloads
This way 1s is the minimum amount of time from the start of one reload
to the next, not from the end of one to the start of the next.

This makes it easier to reason about (without sleeps), but also means we
can now control the ticker in the tests and not depend on flaky tests
that use sleeps.