Do nothing if we receive an unknown redis message
Don't assume it's a reload and throw a warning.

This was very likely the cause of a hot reload test flake:

	$ go test -count 30
	--- FAIL: TestHotReloadSingle (0.00s)
		api_test.go:692: reload wasn't queued
	FAIL

This was because something was firing an unexpected reload that wasn't
being given a tick, meaning that the queue was unexpectedly left full
and thus TestHotReloadSingle failed to enqueue.

This is a potential fix and I can no longer reproduce the flake, even
with -count values of 100.

In any case, the new code is better.