Fix API reloads and add related tests
It broke because reload functionality was using DefaultServeMux for
routing, but our initial `listen` logic used custom muxer, which is
created automatically if you pass non nil second argument to
`http.Serve`.

Change which broke this functionality (#596) was intended to fix
`panic` caused by multiple registrations of “/“ route handle #531.

We did not noticed this bug because we had no tests covering API
reloads, and this PR adds them as well.

This PR fixes API reloads, and still ensure that no multiple route
registrations will be made. It is achieved by creating new
`newServeMux` and updating `http.DefaultServeMux` with it.

Note that our routing and reload functionality have a lot of
duplication (in 3 places), and should be unified one day. Goal of this
PR is to fix the bug, and not to make refactoring.