Remove unnecessary commandModeOptions if statement
v is of type interface{}. The first if triggers if it's of type bool and
value true (non-nil). The second triggers if it's non-nil and not of
type bool and value false.

The second covers the case that would fire the first one - a value of
type bool and value true - so no need for the first.

Moreover, the two have exact bodies.