all: simplify some code
In apidef, we don't need to guard against a nil slice to call len on it.
Moreover, the other methods don't guard against a nil receiver so Len()
shouldn't either. We also don't need to grab the lock just to check for
i < 0.

And in the rpc code, simply use ok == true. If ok is nil, that condition
will be false.

And a couple of small cleanups in the redis code.