hostchecker: remove ID field
It seems to be used to uniquely identify hosts to check. But it's always
just CheckURL. To avoid confusion, remove it and always use that
instead. If an ID of some sort is ever actually needed, it can be added
back.

This makes the fix for #678 trivial. Since the second field is gone,
just remove it from the log call.

Fixes #678.