Replace lonelycode/go-uuid with satori/go.uuid
Another outdated uuid library. Unify with the main one we use.

go-uuid's NewUUID actually returns v1 instead of v4, but that was simply
a misunderstanding - when the library was initially added, the intention
was to use v4 like in the rest of the codebase.

The old String method behaves just like the new one. The format and
length of the strings is the same.

Fixes #524.