proxy: overwrite DialContext, not Dial
Since context was added to Go 1.7, the Dialer type got a DialContext
method and Dial was deprecated.

We don't care much since we don't use the context ourselves, but it's
important to note that if both are set, DialContext takes priority since
the other is deprecated.

Overwrite the method that takes priority.

Found by staticcheck; fixes one of its warnings.