jsvm: clarify js core (tyk.js) loading code
The old Init method took an argument, but discarded it. Make the
situation clearer by not having an argument at all. The parameter isn't
useful for testing either as it's a path to a file.

Add a test that keeps the VM running forever

This is failing until we add a timeout to kill VM copies. Then the test
can simply use a very short timeout to kill the endless loop instantly.

Add a timeout and use it to kill the VM

Unfortunately, the only way to stop/kill Otto is via a panic, which
makes this code more complicated than it should be.

Make the Timeout be 5 seconds for now. It's high enough that no sane
middleware would hit it, but low enough to not let goroutine leaks build
up easily.

Fixes #496.