api: bring back func to list request addresses
In 84417ac2 I introduced a cleaned up GetIPFromRequest and used it
everywhere. I didn't notice the fact that getIPHelper didn't get a
single IP, instead it listed all known request addresses.

Bring it back in a simpler form and use it for all the logs, since they
used getIPHelper before.

Also remove the SplitHostPort of r.RemoteAddr from the recovered version
of the code. Since these are for logging reasons where information is
useful and we don't remove ports from X-Forwarded-For, it's better this
way.