api: rework code that filters by hostname
Make the routers filter by hostname in loadApps and loadAPIEndpoints.
The former has to filter by HostName if non-empty, and the latter has to
filter by either HostName or ControlAPIHostname, the latter taking
priority if both are set.

Do it in the funcs that actually add handlers to the routers as they are
called from many places, and this way we can be sure that the hostnames
are always set up correctly and they only affect the sub-router they
should be affecting.

In the case of the Control API, we were using getHostName() which just
used HostName, ignoring ControlAPIHostname. This resulted in
ControlAPIHostname not being used properly.

Fixes #670.