mw: fix whitelisting of ips with ports, again
In the previous commit, I forgot to fix the Split call to work on
remoteIPString, not r.RemoteAddr.

Before this commit, if X-Forwarded-For was set but r.RemoteAddr
contained a ':', we would use r.RemoteAddr instead of X-Forwarded-For.

Spotted by Leonid during code review of the previous PR, but after it
was merged.

Updates #704.