api: add "block" url parameter to /tyk/reload
If set to "true", it will make the endpoint block until the reload is
done.

Add a test for it, and make TestListener use it. Now that it properly
waits for the reloads to finish, we can get rid of the hacky doReload
and afterFn.

Note that it only blocks on reloads that are queued. If a reload was
already queued, it won't block yet. This will be done in a separate
commit.

Updates #703.