hc: deflake test by controlling the sleep ticker
First, don't rely on time.Sleep in the tests. If the machine is slow
enough, even a timeout of seconds can finish and make our tests fail as
we've seen in the "sync: negative WaitGroup counter" panic on Travis.

Second, don't run CheckHost manually. That defies the purpose of using
the GlobalHostChecker altogether. We're not really testing that it works
if we just use its checker directly, completely bypassing its worker
pool and all its surrounding logic.

With the use of a ticker, the test still takes no time as there is zero
sleep:

	--- PASS: TestHostChecker (0.01s)

Fixes #770.