jsvm: fix goroutine leak in the timeout fix
The code was buggy, in how we would end up sending twice to the channel
most of the time. The defer executes if we don't reach the timeout, so
the goroutine would get blocked at the second send.

Instead of trying to fix it by removing the second send, use a buffered
channel and send zero or one times. This way, it's easy to prove that
the goroutine will exit - it can't possibly block when sending zero or
one values to a buffered channel of size one.

Confirmed with a test that makes use of leaktest to check if the leak is
still there. Before the patch, it correctly finds the leaked goroutine:

--- FAIL: TestJSVMProcessLeak (1.04s)
        leaktest.go:90: leaktest: timed out checking goroutines
        leaktest.go:108: leaktest: leaked goroutine: goroutine 58 [chan send]:
                github.com/TykTechnologies/tyk.(*DynamicMiddleware).ProcessRequest.func1.1(0xc4201cfc20)
                        /home/mvdan/go/src/github.com/TykTechnologies/tyk/plugins.go:149 +0x5e
                github.com/TykTechnologies/tyk.(*DynamicMiddleware).ProcessRequest.func1(0xc4201cfc20, 0xc420a4a950, 0xd734d4, 0x7, 0xc42022a120, 0xd8, 0x112, 0xc4203b2500, 0x222, 0x24a)
                        /home/mvdan/go/src/github.com/TykTechnologies/tyk/plugins.go:153 +0x1a3
                created by github.com/TykTechnologies/tyk.(*DynamicMiddleware).ProcessRequest
                        /home/mvdan/go/src/github.com/TykTechnologies/tyk/plugins.go:140 +0xd69

Unfortunately, we can't include this test as our TestMain starts
goroutines and thus leaktest won't be reliable, especially on CI.

Fixes #804.