test: use httptest.NewRequest for mock requests
But do use http.NewRequest for the ones that do get sent to a real HTTP
server, in this case our local version of httpbin.

This is the proper thing to do, as http.NewRequest isn't suitable for
passing directly to an http.Handler as per the httptest docs. Likewise,
httptest is not suitable for creating real requests that will go through
a real HTTP server.

While at it, add client.Do error checking to testHttp which was helpful
in debugging this.