virtendp: re-add config_data handling
I removed the old GetConfig method here since ProcessRequest made no use
of the config parameter.

What I missed is that GetConfig was actually used in ServeHTTPForCache.

Re-add the code. And use a separate method name to clarify that it's not
the same GetConfig method seen in other middlewares. This way we also
make sure it won't be removed by mistake in the future, as the new
configData method is not part of the base middleware type.