hostcheck: fix case when round-robin host is down
The balancer uses round-robin to select the host to use. If the host is
down, the current behaviour is to try again N-1 times, where N is the
length of all the target hosts.

This works if there are only sequential requests to an API - since
round-robin loops around, trying N times will result in the load
balancer checking if any of the hosts is up, and using the first
relative to where the round-robin first started.

But with many concurrent requests, it doesn't result in checking if all
hosts are up. Since the round-robin state is shared for all requests to
an API, it will be as good as trying a random host N times.

In the case of our test with two hosts and only one being up, this can
result in the balancer checking if only one of the hosts is up twice.
This gets worse as the number of concurrent requests increases.

First, modify TestHostChecker to do a number of concurrent requests. 10
is high enough to reproduce the issue almost every time on my machine,
but low enough to keep the test around 100ms.

Second, fix the issue by doing our own round-robin when the first host
we try is down. This ensures that the balancer will try all hosts for
every request.

Fixes #825.