api: force gorilla/mux to ignore trailing slashes
So that we can easily support /foo/bar and /foo/bar/ for all static
endpoints without having to define them twice nor redirecting one to the
other (e.g. via mux's StrictSlash, which does redirects).

Do this by using {_:/?}, which matches the regex /? and discards it as
a path argument.