api: don't copy request to get an URL parameter
This is when getting the key to find the name of the API version to use.

FormValue did the trick because on top of POST and PUT body forms it
also uses URL query parameters. But that may also parses the body, hence
why we needed the copy.

But all we want is URL parameters, so use that directly and avoid the
request copy.