jsvm: add rawlog func to print directly to log.Out
Without going through logrus and the pretty formatter. This means that
there will be no timestamp and no labels, and that your string won't be
escaped at all.

Call it raw to describe those, and also to have a scary name so that
people wishing to use it are warned that they should know what they're
doing.

Don't backport the test, as it depends on changes made in master.

Fixes #844.