analytics: fix GeoIPDB field race
The GeoIPDB field is initialised in a goroutine but it's read by
multiple goroutines, including the ones from incoming HTTP requests.

There is no reason to call reloadDB in a goroutine - all it does is call
maxminddb.Open, and that func only mmaps the file. It should not take
any noticeable amount of time.

We could also keep it in a goroutine and introduce a mutex, but that
seems overkill for no reason.

While at it, also inline the function and remove its "reloading" logic.
This is because Init is only called once at the top of main in
initialiseSystem, thus we don't actually support replacing the GeoIP
database file at run-time without restarting the gateway entirely.

If we ever want to do that, we should do it properly, including tests
and the data race that we had in master.

This takes care of the following "go test -run=- -race" report:

	WARNING: DATA RACE
	Read at 0x0000018aa340 by main goroutine:
	  github.com/TykTechnologies/tyk.TestMain()
	      /home/mvdan/go/src/github.com/TykTechnologies/tyk/gateway_test.go:153 +0x3dd
	  main.main()
	      github.com/TykTechnologies/tyk/_test/_testmain.go:298 +0x1cd

	Previous write at 0x0000018aa340 by goroutine 18:
	  [failed to restore the stack]

Updates #363.