test: use apisMu and policiesMu in the tests too
And move the tests that ensure the locks are used properly. Instead of
them being separate tests, have the test program continuously write to
the maps in the background. If any test is racy - even if it's in the
test code - then -race will catch it.

This should also make these races show up faster.

And leave the background loop with a 10ms sleep to not spin the CPU too
much. I tried with a runtime.Gosched but it still bumped the test time
from ~2s to upwards of ~3s.

Fixes #891.