all: replace def RawData with simpler ConfigData
Before this commit, RawData was an entire copy of the APIDefinition as a
map[string]interface{}, all the way down. This was only used in a few
middlewares to obtain certain structured data via mapstructure.

In practice, this was only used to obtain the config_data field of type
map[string]interface{}.

Remove the entire copy of APIDefinition in memory and just add a
ConfigData field of type map[string]interface{}. This way, we can also
get rid of all the code that was gluing RawData to its final purpose.

This is in line with the changes being made in other repos to add
config_data to the API definition schema. We don't want to allow
arbitrary root fields - we only want to allow config_data, at least for
now. This makes sure that the gateway doesn't support anything else
unless we meant to.