loader: simplify method and var names
The name of the type is APIDefinitionLoader, so there's no need for the
methods to have a LoadDefinitions prefix. Also, make the method without
suffix be FromDir, as that's descriptive of what it actually does.

Also make the receivers not be pointers - it's unnecessary, as the type
is an empty struct.

Finally, remove APILoader as a global - it's completely unnecessary.