api: redesign round-robin with sync/atomic
This means we no longer use a mutex around a number. The code should be
a bit faster, but most importantly it's much shorter and easier to
understand.

The only change in behavior is that when the max gets shortened, instead
of resetting back to 0, we wrap around with the remainder operation.
This is less predictable - the new position will be random - but it's
actually better. If we always reset to 0, the first elements of the list
would tend to get selected more often than the last elements.

Also test that we don't barf with a divide by zero if the length is zero
and there are no positions to select from.