all: don't make the global api spec map a pointer
It's unnecessary, as maps are already pointers. This was done in master,
but not yet in stable.

The tricky part is that the double-pointer means that reads can't be
safely done without checking for nil first, which can be done if the
double pointer is removed.

This was missed in the backport of the API deduplication feature, which
led to a nil pointer dereference crash.

Fixes #938.