all: always make mainRouter == defaultRouter
setupGlobals sets them up to be equal, but they might diverge under some
circumstances like a reload.

In some cases we register things on defaultRouter instead of mainRouter.
If they're not the same pointers, things start breaking.

This was the case with OverrideDefaults.

Don't merge the two global vars in stable, as that change is a bit too
invasive. Simply fix the two occurences where they can diverge.

Fixes #940.