mw: avoid double goroutine in rate limiting
If UseAsyncSessionWrite is true, DefaultSessionManager.UpdateSession
does the actual store write in a goroutine. Other than that, the
function has very little logic and a json marshal.

As such, it's redundant for this middleware to also use a goroutine. It
would only lead to extra CPU use and delay in writing the session to the
store, given how two goroutines need to be scheduled instead of just
one.