jsvm: regression test for non-string config_data
After the fix in stable. master was fixed by accident, so add a test to
make sure we don't break it - again - by accident.

Use the virtual endpoint test as it goes from a JSON string, to make
sure the nested object and non-string types make it through JSON
decoding too.