revproxy: don't append paths if we switched URLs
Because if we did, we may end up with a duplicate path.

We already partially fixed this with a path string comparison. However,
after some debugging, it was discovered that under some circumstances,
paths may be of the form "path" instead of "/path".

Use a stronger logic, and add tests for all combinations of RetainHost
and unexpected paths.

Reproducing the issue manually, before and after:

        DEBUG Outbound Request: http://192.168.0.148:8888/_count/_count

        DEBUG Outbound Request: http://192.168.0.148:8888/_count

Also move the scheme and host swapping, as they too modify req.URL when
it's already final. Although these were harmless, as they were replacing
instead of appending.

Fixes #962.