Add a new feature to add headers from a body payload
This allow to take some data from the json payload during a JQ
transformation, and put it in the request/response HTTPS headers.

To enable this feature, it's needed to update your JQ transformation
filter with a new JSON object with the "output_headers" field which
contains a objects with the header name/ header value pair.

Example:

. as [$context, $body] | $body, . as [$context, $body] | {"output_headers": {"DEVICEID": "CSE3219/C9886: $body.DEVICEID}}"