rpc: make connection ID a local variable
This value is only used in its block, so it need not be a global. But
the problem is that we may call Connect multiple times to force a
re-connect, and the global was getting overwritten while a Dial call was
reading it.

This is the data race that CI caught. Fix it by just making it a local
variable. That way, it's impossible for two distinct Connect calls to
share the same connection ID variable, thus there is no possibility of a
race.

Fixes #987.