session: undo MetaData map[string]string switch
This was seen as a safe change because the gateway never uses values
other than strings, nor does it have tests that use it any other way.

However, the dashboard does use nested objects with strings in them.
Thus this change broke the dashboard.

Revert the change, updating the old map conversion TODO and adding
another one in the fmt.Sprintf call, since we probably want something
stricter and more predictable.

There is still work to do here, as this still allows for too many types.
However, reverting is the right decision now to un-break the dashboard
for 2.4.