test: use a random port for the mock rpc server
We tried this before, but failed because we weren't waiting for the
server to start up, thus we didn't have the final address.

Now that we do, this is fairly trivial. This should get rid of the
"address in use" flake on CI. It's not clear what is causing the flake,
as we haven't been able to reproduce it locally, but there is no need to
use a fixed port in any case.

Fixes #835.