loader: remove some temporary mw chains
These all get almost immediately appended to the main chain and don't
have any other use. They make the code especially hard to read because
it's hard to tell the final order of the middlewares, since you have to
keep track of multiple chains and when they're merged.

While at it, unexport the middleware chain helpers.