mw: simplify and improve size limit mw
Join the two copies of the same code into one by merging the size
numbers instead.

While at it, do the calculations as int64 instead of int, to make sure
we don't accidentally break the numbers. If we support 64 bits for the
limit, we should do the same for the incoming request size.