all: remove a bunch of ReadFile/ReadAll calls
The standard library and plenty of libraries tend to work with
io.Readers as well as string/[]byte, so most of the time we don't need
to read entire bodies and files into memory.

This also simplifies the code a bit, since we remove an extra step in
some cases.

While at it, stop ignoring some unmarshalling and parsing errors that
were spotted while refactoring.