all: simplify redis manager object logic
First, get rid of a shared object in the webhook handler code. All redis
managers share a redis client singleton under the hood nowadays, so it's
completely unnecessary. But most of all, that code is racy.

While at it, inline a couple of helpers that were used in only one
place.