all: simplify some string handling
Use prefix and suffix matches instead of Contains when that's what we
want.

Also refactor some manual string handling to get the first part of a URL
path. It is much clearer, as well as more foolproof, to use
strings.Split and a loop. For example, the old code worked with /v1/foo
but not with //v1/foo - and of course was more complex.