dash: don't force a reload after a relogin
doReload just does an actual reload. It should never be used in
production code, as it's only meant for the reload loop. If it's used
directly, we might end up with two reloads happening concurrently.

If that happens, we can run into all kinds of race conditions.

Simply queue a reload and wait for it to finish.