remove CHANGELOG.md from the root of the repo
This file used to hold the changes in each version. Since early this
year, we've switched to placing the changelog information on the GitHub
releases page, and we stopped updating this file.

We decided to keep it around for a little bit, removing it once it grew
too outdated. The time has come - it's eight versions behind.