mw: fix transform middleware regression with XML
This was a crash in the mxj library concerning invalid input from a
reader, as it would not properly error and stop. We did not hit this
before in 2.3.x as we used a []byte, not an io.Reader.

The upstream issue is https://github.com/clbanning/mxj/issues/38, fixed
back in February.

Add a regression test and bump the vendored version.

Fixes #1122.