proxy: fix and test multi-target reverse proxy
I broke it when replacing the false Init method with a regular
constructor, since I forgot to replace the nil return with the actual
object.

Add a test that covers both the default proxy and an overriding one. It
was crashing as expected. Fix the constructor, making the test pass.

While at it, clean up the code quite a bit:

* no need to parse TargetURL again
* getProxyForRequest can be much simpler
* unindent some of the code

Fixes #1177.