loader: simplify auth chain construction
First, move the "enabled" checks to EnabledForSpec, which is why it's
there in the interface to begin with. Let the loader still log when a
middleware was appended by making mwAppendEnabled return a bool.

Second, remove a duplicate useCoProcessAuth if statement.

Third, greatly simplify the logic to decide whether or not to use
AuthKey. The second part of the conditional checked that no other auth
middleware was loaded. We can do this by simply checking if the auth
chain is empty, instead of using a very complex and hard to maintain
chain of bools.