all: clean up and simplify a bunch of code
First, we know that in the versions map, the key will always be equal to
value.Name. The docs don't explicitly state this, but other parts of the
code such as elsewhere in the gateway and the dashboard use the two
interchangeably.

Do the same in the API definition loader. It was caching both the key
and value in the request's context. Cache only the value. We only ever
needed to cache the version key here, and it won't be useful in the
future, so remove the machinery too.

While at it, collapse some identical switch cases and make
ParseDefinition take a reader, since a string is not necessary.

Finally, remove a test API definition file I recently introduced by
mistake.