config: expose default config as-is
Since it can be useful to other packages. Unexport WriteDefault, as
that's the hacky way we were accessing the default config instead.

While at it, also give the Monitor config field a proper json key name.
This was an oversight introduced ages ago, when the config as it is now
was introduced. This is a backwards compatible change, given that
encoding/json is case insensitive when unmarshaling. So "Monitor" on
people's JSON configs will keep on working as it was.

But marshaling of configs will use the consistent key name, which is
what we're after.