events: move some type assertions into Init funcs
Some fields in the handlers were of type map[strint]interface{}. We only
ever used one value from that map, having to type assert it at each
event that was handled.

This is overly complex and wasteful. Fetch the value once at Init,
reducing the complexity and amount of work in HandleEvent, which will be
called many more times.