storage: a couple of redis cleanups
First, don't make all the methods use pointer receivers. None of them
modify the struct, and the struct itself is tiny - just a string and two
bools, which should total three ints. This should mean a slight memory
and performance win, as we don't have to allocate as many objects and we
don't have to jump around memory as much.

Second, refactor GetRelevantClusterReference into a method with a
shorter name, making the code more understandable and simpler.

While replacing RedisCluster pointers with structs, I also noticed how
RedisNotificationHandler could discard notifications if the store wasn't
initialised. Instead, synchronously set up the store, so we know it has
been set up. This is not a problem, given that Connect doesn't actually
block connecting to all the hosts.