api: use a strip slashes middleware
To end the gorilla/mux trailing slash madness. Do this only for the
admin API, where we were using this workaround.

While at it, remove an unnecessary subrouter, improve a test error
message, and test that nested net/http/pprof endpoints work.

Note that a new keys route was added, as the existing one requires a
slash. Before, we only supported /keys/ and /keys/foo, but not /keys.
Now all three are supported. This surfaced now given how we rewrite the
/keys/ path into /keys.