Make HTTP transport API spec property
Right now Transport is created for each requests, which means that
connection pooling is not used and on each request it creates new
connection to the upstream server.

This is differ from 2.3 behavior, which had global Transport for all
APIs. During the 2.4 refactor it was changed to behavior described
above.

This change introduce transport caching, so it gets re-used between API
calls. In addition, this is better than 2.3 approach, because each API
can have different transport settings and timeouts, and now each API
gets own transport.