test: add an api loader sort test
Wrote it while investigating #1271. The test passes, so it's unlikely
that the sort itself isn't broken, but this test is useful nevertheless.

It loads a bunch of API definitions with overlapping listen paths in a
semi-random initial order, and then uses them all making sure that the
right API spec responded to the request.