all: fix numerous global apiSpecs data races
Also make gateway_test.go help the race detector by writing to it in the
background. This made the races show up on every 'go test -race' run on
my laptop, as opposed to just sometimes.

While at it, don't use apiSpecs as a name for local variables, as that
makes the code very confusing.