Packaging fixes and improvements
This adds a "tyk" user and group for DEB and RPM packages so that package files and directories would be owned by it and the process run with its effective uid and gid. Appropriate changes made to build scripts and init configs. In addition to this PID now has to reside in its own sub-rundir due to this change, so that's created (and additionally managed by systemd where it's available), default pidfile location changed appropriately so that upgrade wouldn't require any config changes by the users.

The gateway config file is now only readable and writable by the "tyk" user/group.

The build and push scripts allow manually assigning a version based on an environment variable, which is useful for unstable packaging (e.g. based on git revision ID).

Doesn't remove "default" init system files on upgrade/remove anymore as it's now a way for users to run the process with custom environment variables, hence sort of a config file.

Does a bunch of fixes on the packaging and init scripts (e.g. fixes RPM upgrades removing init scripts on upgrade). Removes old stuff. Adds Ubuntu 16.04 LTS (Xenial) package to PackageCloud.