cache: EnabledForSpec should just use EnableCache
Don't require there to be a cached extended path too. Otherwise, one
can't use any sort of caching on an API without adding cached extended
paths.

At the same time, don't run the middleware if there are any cached
extended paths yet EnableCache is false. The docs tell the user to
enable the cache for the whole API, so that is expected.

Add a test to prove that caching of a simple API works. This test was
failing before, as the API does not have any cached extended paths.

Fixes #1329.