bug: implement EnabledForSpec for TrackEndpointMiddleware
fixes: #1547

Because `EnabledForSpec` was not implemented for this middleware, the mw
would run for every request - even if it didn't need to.

Implementation of the method resolves this issue. When analytics are
disabled or api spec is set to do_not_track, the middleware will
no longer be loaded. This should also result in a slightly faster response
time.