PoC - support yaml and json config files
This is just a PoC - Not for merging, and obviously needs to be fully
tested.

json is difficult for OPS teams to work with, especially via the command
line.

without going into too much detail about benefits of yaml, we can also
add comments to configs, or even comment some fields out. even offer
explanations / documentation directly into the config file if necessary.

json is valid yaml. And as such, with a few simple changes, instead of
unmarshal json, we can unmarshal yaml using the github.com/ghodss/yaml
lib.

if data is json, it will unmarshal into struct. if data is yaml, it will
also unmarshal into a struct.

The benefit of this method, is that we can support both yaml and json
configs without having to worry about adding extra struct tags.

Further - if this works well, we could potentially also support yaml api
definitions, and maybe even yaml swagger specs with little effort.