coprocess_python: refactor the dispatcher to re-use the hook table for event handlers, remove unused code
Previously we had separate data structures for "request hooks" and "event handlers". In the current approach we require an entrypoint file ("middleware.py") to set all the functions, because of this is much simpler to merge the logic of both event handlers and request hooks.
One of the issues that this brings up is that event handler names and request hook names need to be unique in the same plugin context but this shouldn't be a major issue.
Also applies changes related to #1338.