Handles Dashboard API failures, fixes #1725 (#1726)
Have added a test that simulates a running gateway with 1 definition, and then that same gateway pulling a malformed definition

The gateway will default to what is already loaded.

This introduces the potential for gateways to get out-of-step but ensures existing services stay loaded until there has been a successful reload.

Perhaps a future improvement would be a signal to the dashboard notifying which gateways have successfully reloaded