Reuse client for heartbeat. (#1946)
Fixes #1951
This prevents the issue with spawning a new open socket for each gateway to dashboard heartbeat that is observed in TLS connections when the default transport is overidden.

From Go client.go:

```
// The Client's Transport typically has internal state (cached TCP
// connections), so Clients should be reused instead of created as
// needed. Clients are safe for concurrent use by multiple goroutines.
```

Since we are doing the same operation constantly it makes more sense to create a single client and reuse the connection for the dashboard heartbeat.