Add special handling for `context canceled` error
In this specific case Gateway itself did the right job and proxied the request, it is the client who messes with the connection. This is still an error state, but there should be a way to differentiate such issues from server-side ones. In this cases we introduce special 499 code (same as Nginx has), which description is:

```
499 Client Closed Request
Used when the client has closed the request before the server could send a response.
```