Fix handling of broken HTTP clients (#1979)
Clients who close connection before reading response, cause 500 error in gateway analytics

In this specific case Gateway itself did the right job and proxied the request, it is the client who messes with the connection. This is still an error state, but there should be a way to differentiate such issues from server-side ones. In this cases we introduce special 499 code (same as Nginx has), which description is:

```
499 Client Closed Request
Used when the client has closed the request before the server could send a response.
```

Fix https://github.com/TykTechnologies/tyk/issues/1980