Fix broken logstash UDP support (#2141)
This PR disables `gas` reconnection functionality when `"logstash_transport": "udp"` because it doesn't support `udp`.
 
Also, we need to fix doc here: https://tyk.io/docs/log-data/#a-name-aggregated-logs-with-logstash-a-aggregated-logs-with-logstash. It says `logstash_transport: The Logstash transport to use, should be "tcp".` but we support `udp` too. And, it would be easier for user if we give a logstash input configuration there like:

```
input {
	tcp {	
		port => 5044
		codec => json
	}
}
```
Fixes #2136