res_handler_transform: make sure bodyData is available on XML errors (#2133)
For #2016.

I was able to reproduce using the instructions described in the ticket. The issue occurs at this point:

```go
	bodyData := make(map[string]interface{})
	switch tmeta.TemplateData.Input {
	case apidef.RequestXML:
		if len(body) == 0 {
			body = []byte("<_/>")
		}

		mxj.XmlCharsetReader = WrappedCharsetReader
		var err error

		bodyData, err := mxj.NewMapXml(body)
		if err != nil {
			logger.WithError(err).Error("Error unmarshalling XML")
		}
```

When `mxj.NewMapXml` returns an error, `bodyData` is not properly set and further references to it fail (like setting context vars using `bodyData["_tyk_context"] = ctxGetData(req)`).

This approach avoids replacing `bodyData`.