Bug/2195 fix looping issue (#2200)
Fixes #2195

Every middleware call CheckSpecMathesStatus() to check if middlewares are set for the path.

In case of url-rewrite, path set in ctxGetUrlRewritePath was used, which stores value of original path before url-write. That variable was used to run response middleware of the original path and not of the rewritten path

But in case of looping feature, we want to run request middlewares of the rewritten path as well. So I changed the code to use ctxGetUrlRewritePath only in case of response middleware and use r.url.Path otherwise.