Update accessRights of Per API policy when API is deleted (#2312)
Fixes [tyk-analytics/issues/1222](https://github.com/TykTechnologies/tyk-analytics/issues/1222)

`accessRights` are stored in a map. While applying Per API policy, previous instance of map was used and entries where updated. So if an API was deleted from policy, it's entry was not deleted from `accessRights`. 

The issue is resolved by creating new instance of map and adding entries to it.