Unauthorized vs Forbidden
http status unauthorized - 401 - token is invalid, i don't know who you are.
http status forbidden - 403 - token is valid, i know who you are, but you are not allowed to access this resource.

By the time we are applying policies to the session token, this means we know who the client is, and are validating whether they sholud be able to access a resource or not.