Extending advanced caching to support all of the methods, adding CacheOnlyResponseCodes per method and the ability to use headers as part of the cache key (#2753 #2763) (#2754)
Example of a use-case:
We want to cache put/post methods that receive 401 or 403.
Next time the same request is received we will return the response from the cache unless they change the authorization or tenant headers.

This PR contains:
- Cache support for all of the methods, it can be set from the endpoint designer
- Each cache endpoint can have its own list of ResponseCodes that will override the main API cache list for this endpoint
- Added a list of headers per endpoint that will be inserted (with header values) into the cache key
- Some refactoring to reuse some methods
- Tests

More info can be found in the feature requests #2753 and #2763

Closes #2753
Closes #2763