Fix memory leak (#2879)
We stored HTTP handler inside API definition, which probably introduced some circular dependency, confusing the GC.
Handler access itself was needed for the looping, so even if API not loaded to the router (internal), it still could be accessed.

Now handler access moved to the map, where key is API id.
And map is updated in the same maneer as apisByID