Oauth clients and certificates listing optimization (#2984)
The idea behind this PR is solve the slow loading of oauth clients and certificates lists.

This is caused because we scan all the redis keys looking for the certificates / oauth clients.

This PR solves this creating an index list per org (for certificates) and per API (for oauth clients) where the keys are going to be migrated.

The behavior expected is the first time we load the list, it's going to be like it is now but all the keys are going to be migrated for that certs list / oauth client (if they already exists). In the second load, it's going to be faster,  since it get the keys from the list.