Fix reload  implementation
This change removes  `startReloadChan` and `reloadDoneChan` because
they are reduntant and blocky. Under high concurrecny they result in the
reload loop to block resultinto new policy changes not be reflected on
the gw.

Also, `reloadQueueLoop` is siimplified and do as what the name implies,
only queuing callbacks to be triggered on the next reload tick.

This also make sure both `reloadQueueLoop` and  `reloadLoop` are
cancellable by accepting `context.Context`