Fix reload test timeouts
Majority of the tests for reloads were operating under false assumptions
The ReloadTicker will  immediately trigger a reload, however queuing
happens in a separate goroutine so its cricial to synchronize the
operations to make sure reload happens.

This change adds ReloadTestCase, which exposes methods to work with
reloads in tests. This covers all cases of ensuring queue happened and
also ensuring reloads happened.