Fix TT-1311
Fixed slow certificate enddpoint for orgs who have no certs.

The main root cause of the issue, is that initially when we implemented certificates we added a performance issue because it use Redis Scan command to list certs, and it is slow blocking operation. After we realised that on cloud it is very slow, we added a fix, so no we maintain a special collection with “certificate ids”, to make it fit into redis model from performance point. Together with this change we also added a “auto” migration for such users, so if you do not have this “special” redis collection, it will create it for you, and will perform initial “slow” scan to fill it.  For users who have no certificates, it can’t find such “special” collection, and it tries to perform this migration. It can’t find any certificate, and does not create this record.

Redis collection gets removed if there are no records.
So I added a new key, saying that migration is done.

https://tyktech.atlassian.net/browse/TT-1131