Fix TT-1311 (#3410)
Fixed slow certificate endpoint for orgs who have no certs.

The main root cause of the issue, is that initially when we implemented certificates we added a performance issue because it uses the Redis Scan command to list certs, and it is a slow blocking operation. After we realized that on the cloud it is very slow, we added a fix, so no we maintain a special collection with “certificate ids”, to make it fit into Redis model from the performance point. Together with this change we also added a “auto” migration for such users, so if you do not have this “special” redis collection, it will create it for you and will perform an initial “slow” scan to fill it.  For users who have no certificates, it can’t find such a “special” collection, and it tries to perform this migration. It can’t find any certificate and does not create this record.

Redis collection gets removed if there are no records.
So I added a new key, saying that migration is done.

https://tyktech.atlassian.net/browse/TT-1131

Some pages on legacy cloud became unusable for users who have no certificates.

- [x] Bug fix (non-breaking change which fixes an issue)

- [x] Make sure you are requesting to **pull a topic/feature/bugfix branch** (right side). If pulling from your own
      fork, don't request your `master`!
- [x] Make sure you are making a pull request against the **`master` branch** (left side). Also, you should start
      *your branch* off *our latest `master`*.
- [x] Modules and vendor dependencies have been updated; run `go mod tidy && go mod vendor`
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [x] Check your code additions will not fail linting checks:
  - [x] `go fmt -s`
  - [x] `go vet`