Goreleaser/all tags (#3477)
# Description
These related PRs fix the integration not pushing all tags. This is an urgent fix needed for QA.

It improves the execution by splitting up the jobs.

# Testing
The status of the integration workflows can be inspected in this PR itself in the _Checks_ tab.

# Merging
Once this PR is merged, it will be automatically sync'd to release-3-lts, release-3.2 and master.