Tweak publishing strategy (#3482)
# Description
These related PRs fix the implementation in line with https://tyktech.atlassian.net/wiki/spaces/DEV/pages/449708061/Release+Engineering#Publishing

When goreleaser detects that the current commit is not a tag, it goes into snapshot mode. Artefacts created in snapshot mode are not uploaded anywhere. 
When a -qa or -rc tag is pushed, the deb and rpm files are published to the -unstable repo and the docker image tagged unstable is updated.
When an v<semver> tag is pushed, deb and rpm files are published to the stable repos and docker image is pushed to Docker Hub tagged with the pushed version and v{{Major}}{{Minor}}.
When a tag is pushed, goreleaser goes into release mode and will,
-draft a github release with a generated changelog
-attach tarballs, deb, rpm to the above release
-publish all docker images to Docker hub
-publish deb and rpm files to the stable repository

tyk-pump already saw some of these fixes in https://github.com/TykTechnologies/tyk-pump/pull/316 so the diff may look different there.

nfpm demands semver-compliant tags so they are mandatory now.

tyk-sink uploads were failing due to the wrong repo name. 

# Testing
The status of the integration workflows can be inspected in this PR itself in the _Checks_ tab.

# Merging
Once this PR is merged, it will be automatically sync'd to release-3-lts, release-3.2 and master.