Work with subrouters inside processSpec (#3486)
The master router is passed to the `processSpec` and while adding handlers to it always needs to put listen path in front. I realized that it is more than a redundancy. For example, if I want to add CORS middleware to subrouter, it wasn't matching with the listen path because they were living in the master router. This PR gets subrouter first and passes it to the `processSpec` function.

- I saw that there is a problem in graphql playground. So I created a PR to that repo as well: https://github.com/TykTechnologies/graphql-go-tools/pull/100

I created this PR as draft, after merging graphql PR, I will update `go.mod` and this PR will be ready.

(cherry picked from commit 8fc3b6cff78d148d598ff359c99e7c9f170dcd8a)