test(gateway): fix flaky test TestGraphQL_HeadersInjection
Expecting map elements to come back in a specific sequence is courting failure. Asserting with
strings.Contains will give a more reliable result. Also, switched the defer call in this test to use
t.Cleanup as the testing.T struct is closer and more appropriate to the control flow of running
tests.

fix TT-2021