releng: release-3 is for all future 3.x branches (#3557)
Until now from a release engineering perspective, release-3 was intended as a backport branch to backport features into LTS releases. This PR brings release engineering from master to release-3, making it ready for any future branch to be cut from it. Henceforth, there will be no automatic sync originating from release-3.

Release engineering on the following branches will need to be fixed by
hand:
- release-3.1
- release-3.1.0
- release-3.1.1
- release-3.1.1-cloud
- release-3.1.2

And the same thing will have to be done in tyk-analytics.

## Related Issue
https://tyktech.atlassian.net/browse/TD-341
https://tyktech.atlassian.net/browse/TD-340

## How This Has Been Tested
Results in the _Checks_ tab.