Reduce API mutex locking
At the moment it locks for the whole period of getting APIs from Dash/RPC and parsing the output.
In the context of Legacy Cloud it can be a few seconds.

Mutex should be locked only when we swap the global variable.

Potentially shoud fix https://tyktech.atlassian.net/browse/TD-123