[TT-2932] fix upgrades for gw from 3.0.6 (#3630)
This is an alternate strategy to https://github.com/TykTechnologies/tyk-analytics/pull/2386

## Description
With this PR, we fallback to the old Buddy pipeline release process for 3.0.x releases. This PR will need to be applied to release-3.0.7 and release-3-lts.

## Related Issue
https://tyktech.atlassian.net/browse/TT-2932

## Motivation and Context
We backup and restore the systemd service file using rpm specific triggers. Since goreleaser and xenial are in opposition, using fpm can allow for LTS releases to continue supporting xenial _and_ work-around the upgrade bug.

While this works there are a lot of scripts that run at various points and it is not easy for a person without context to understand what is going on here. This is not good for LTS branches which prioritise stability.

## How This Has Been Tested
Tested locally using `tyk-gateway-3.0.7~273.3268ac0` from tyk-gateway-unstable repo. Upgraded from 3.0.6 from tyk-gateway repo.