[TT-2932] systemd service restore for deb (#3660)
## Description
This PR fixes the systemd service file being removed on upgrades from 3.0.6 for debian only. The RPM version of this fix has already been merged.

This PR is dependant on the the install scripts being templates in tyk-ci/wf-gen. 

## Related Issue
https://tyktech.atlassian.net/browse/TT-2932

## How This Has Been Tested
See the status in the _Checks_ tab.

# Merging
This PR is is for master and future branches only. This should _not_ be cherry-picked to release-3.0.7. It has already been implemented on release-3.0.7 on gw and dash.

For all other repos, please follow the jira above.
<!-- Automatic merging will be performed for the following branches when this PR is merged to master. -->

<!-- | Repo          | Branches                 | -->
<!-- |---------------|--------------------------| -->
<!-- | tyk           | release-3.2              | -->
<!-- | tyk-analytics | release-3.2              | -->
<!-- | tyk-pump      | release-0.8, release-1.0 | -->
<!-- | tyk-sink      | release-1.7, release-1.8 | -->

<!-- Please cherry pick this into the following branches: -->

<!-- | Repo          | Branches                 | -->
<!-- |---------------|--------------------------| -->
<!-- | tyk           | release-3, release-3-lts | -->
<!-- | tyk-analytics | release-3, release-3-lts | -->