fix: update JWT dependency to address CVE
jwt-go allows attackers to bypass intended access restrictions in
situations with []string{} for m["aud"] (which is allowed by the
specification).
Because the type assertion fails, "" is the value of aud.
This is a security problem if the JWT token is presented to a service
that lacks its own audience check.
There is no patch available and users of jwt-go are advised to migrate
to golang-jwt at version 3.2.1.

Seeing as github.com/dgrijalva/jwt-go is no longer maintained, I have
switched to the latest version of the recommended community-maintained
fork instead.