Update README.md
Hello, we are TeamCode. We have created a Tin application for your project, which help your users to quickly run and try TYK without having to install and configure the environment. You don't need to pay for this service. Hope it can help you better promote the project. And bring a better experience to your users.
 you can try it: https://www.teamcode.com/tin/clone?applicationId=268088907964952576
 Read the guide: https://www.teamcode.com/docs/en-US/tin/clone-tin