[TT-4655] Federation crashes gateway during performance tests (#3910)
PR for [TT-4655](https://tyktech.atlassian.net/browse/TT-4655)

Basically, Tyk gateway creates a `memconn.Provider` to handle subgraphs for every request. it also creates a new HTTP client with a fresh `http.Transport`. 

It repeats this flow for every request and the dialer implementation in the Go HTTP library allocates new buffers to dial the upstream server. This creates a massive memory leak in the runtime.  I implemented a simple caching approach to overcome that problem. Now it uses a global `http.Client` instance and the dialer function gets a `memconn.Provider` for the subgraph. Before the fix, Tyk gateway had almost 12GB of memory on my local setup. With the fix, it’s stable at around 140 MB.

Users can rename the subgraph or delete it completely. If a subgraph is renamed, a new `memconn.Provider` will be created and cached.

From memory profiler:

![Screen Shot 2022-02-25 at 12 20 51](https://user-images.githubusercontent.com/33702537/155992064-dde90b72-77bf-45f4-9b75-86f40d0a4cd6.png)