Fix plugin build script.
There's a new module dependency for the gateway that's internal
to the tyk gateway repo. (`github.com/TykTechnologies/tyk/certs`).
This causes the entire gateway source to be deleted in the build
script step where old dependecies are deleted so that they could be
replaced with fresh dependencies vendored in the GW source directory
of the plugin compiler image. This fixes that issue by filtering
the tyk repo from the delete list.
 * Also fixes a variable naming issue(CGOENABLED)