Add public function to directly access the Redis Universal Client
This is needed for plugin developers (in some cases) to issue commands that are not supported by the storage abstraction layer or have otherwise had their behavior changed. For example, it is not possible to issue a `SETEX` command. Similarly, the `MGET` command behaves differently and doesn't work in cases where some keys are expected to be missing.