Codeowners (#4535)
Update/Sync Releng templates for tyk


:robot: Beep, boop, this PR has been generated by
[automation](https://github.com/TykTechnologies/gromit/blob/master/cmd/policy.go)
on Wed Dec 14 19:54:02 UTC 2022

This PR updates the release engineering specific code in the tyk repo,
and has been generated because some releng code was changed, and
the corresponding templates were updated.

These changes will mostly be limited to the `ci/` directory and
`.github` directory.

If you think this was made by mistake, or if merging this might cause
any instabilities,
please log a jira ticket in the TD project citing this PR.

Co-authored-by: Gromit <policy@gromit>
Co-authored-by: Alok G Singh <alok@tyk.io>