[TT-5997] Make MDCB work without group id (#4678)
<!-- Provide a general summary of your changes in the Title above -->

## Description
If a gateway worker does not have an assigned `group_id`, it will be
defaulted to "ungrouped". Note that having more than one ungrouped
worker in different clusters will lead to one of them not receiving any
event.
<!-- Describe your changes in detail -->

## Related Issue

https://tyktech.atlassian.net/browse/TT-5997?atlOrigin=eyJpIjoiZmNkYTI0YTJjZGExNDdlOGI4NDNkZDQ4ZjkxOGEyYzUiLCJwIjoiaiJ9
<!-- This project only accepts pull requests related to open issues. -->
<!-- If suggesting a new feature or change, please discuss it in an
issue first. -->
<!-- If fixing a bug, there should be an issue describing it with steps
to reproduce. -->
<!-- OSS: Please link to the issue here. Tyk: please create/link the
JIRA ticket. -->

## Motivation and Context
In MDCB we have `Login` and `LoginWithGroup` functions. Both of them
have a similar logic, so the idea is to only use one of them. In this
case, it's `LoginWithGroup`, since we're setting a default group to all
the workers without a `group_id`.
The idea of this PR is to make worker gateways to only execute 
<!-- Why is this change required? What problem does it solve? -->

## How This Has Been Tested
It's been manually tested and unit tests were also created.
<!-- Please describe in detail how you tested your changes -->
<!-- Include details of your testing environment, and the tests -->
<!-- you ran to see how your change affects other areas of the code,
etc. -->
<!-- This information is helpful for reviewers and QA. -->

## Screenshots (if appropriate)

## Types of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Refactoring or add test (improvements in base code or adds test
coverage to functionality)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply -->
<!-- If there are no documentation updates required, mark the item as
checked. -->
<!-- Raise up any additional concerns not covered by the checklist. -->

- [ ] I ensured that the documentation is up to date
- [ ] I explained why this PR updates go.mod in detail with reasoning
why it's required
- [ ] I would like a code coverage CI quality gate exception and have
explained why

---------

Co-authored-by: Sredny M <sredny.buitrago@gmail.com>

(cherry picked from commit 4835ec7e8f0b5d3c35f7f66b1308342b24e3f50c)