[TT-7959] Add option to skip ClientCA announcement during mTLS
At the moment we announce ClientCA to the client, which is a default Go way to
do mTLS, and it also gives client a hint which certificate to use, if it has
many.

However if you have a lot of certificates in ClientCA, like 100, it can overflow
TLS handshake buffer of some clients (for example default JVM buffer is 32kb).

This PR adds an option `http_server_options.skip_client_ca_announcement`, which
stops announcing ClientCAs during handshake. However, if you just pass empty
ClientCA, and at the same time set TLS mode to `RequireAndVerifyClientCert` it
will error, because non empty ClientCA required for default mTLS verification,
which is built in to Go TLS library. Which means that we have to "disable"
default client cert validation by setting TLS mode to `RequestClientCert`, and
implement our "own" verification. Thankfully Go provides TLS hook
`VerifyPeerCertificate` where you can add custom validation. It does not have
access to client hello message, and other dynamic variables, so it needs to use
a closure (this way is recommended by Go team). Inside this check duplicate
exactly the same code which Go TLS library has, but instead of depending on
`config.ClientCA` it use certificate pool passed via closure.

mTLS unit test now runs 2 times, with and without clientCA announcement.

I was afraid to make new way "default" because mTLS is weird, and better have
a lot of options to tune it.

But despite new config variable is added, this PR still can count as bugfix.