Sync releng templates.
From the WIP branch fix/td-1528/update-releng-templates
on gromit.
The gromit branch/PR should be merged to master before
this PR gets merged,as sync automation would run on tyk
once this is merged.

For all the versioned releng templates to work with
sync-automation - it needs the template changes that
was used to generate this PR on gromit master.