Add max_content_length, max_request_body_size (#5043)
<!-- Provide a general summary of your changes in the Title above -->

## Description

The PR implements a `max_content_length` addition in the area of
`max_request_body_size`.

The check for `max_content_length` responds with HTTP 411 Length
Required if the content-length header is not present. A handled
exception to this is `Transfer-Encoding: chunked`, which doesn't provide
a content length header and thus can't handle before reading the full
request body. Per-API middleware has access to the body and can validate
length.

The `max_request_body_size` limits the amount of bytes read for the
request if the content-length is unknown or unset. Because there are
exceptions that bypass `max_content_length`, this has a separate flag.
Depending on wanted behaviour, one or the other option may be
configured.

Both flags issue a response with HTTP 413 Entity too big, based on the
content-length header value only.

<!-- Describe your changes in detail -->

## Related Issue

https://tyktech.atlassian.net/browse/TT-5186

<!-- This project only accepts pull requests related to open issues. -->
<!-- If suggesting a new feature or change, please discuss it in an
issue first. -->
<!-- If fixing a bug, there should be an issue describing it with steps
to reproduce. -->
<!-- OSS: Please link to the issue here. Tyk: please create/link the
JIRA ticket. -->

## Motivation and Context

<!-- Why is this change required? What problem does it solve? -->

## How This Has Been Tested

<!-- Please describe in detail how you tested your changes -->
<!-- Include details of your testing environment, and the tests -->
<!-- you ran to see how your change affects other areas of the code,
etc. -->
<!-- This information is helpful for reviewers and QA. -->

## Screenshots (if appropriate)

## Types of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Refactoring or add test (improvements in base code or adds test
coverage to functionality)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply -->
<!-- If there are no documentation updates required, mark the item as
checked. -->
<!-- Raise up any additional concerns not covered by the checklist. -->

- [ ] I ensured that the documentation is up to date
- [ ] I explained why this PR updates go.mod in detail with reasoning
why it's required
- [ ] I would like a code coverage CI quality gate exception and have
explained why

---------

Co-authored-by: Tit Petric <tit@tyk.io>